Removed a source code about a cache_barrier and a profiling
It is for the generality of Dual-Thread Journaling.
Now, the commit thread does not wait for dispatching and
does not dispatch the commit block write command.
And, the flush thread waits for the completion of the write command
dispatched by the commit thread, and then dispatches
the commit block write command with REQ_PREFLUSH and REQ_FUA.
- Joontaek Oh.